Product weight fix

+++ internal/model/product.go
@@ -4,6 +4,6 @@ type Product struct {
 	ID 			string 	`jsonapi:"primary,products"`
 	Name 		string 	`jsonapi:"attr,name"`
 	Description string 	`jsonapi:"attr,description"`
-	Price 		float64 `jsonapi:"attr,price"`
-	Weight		float64 `jsonapi:"attr,weight"`
+	Price 		float32 `jsonapi:"attr,price"`
+	Weight		int 	`jsonapi:"attr,weight"`
 }
